+++ Andriuta/Parser/Transactions.rb
@@ -1,0 +1,9 @@
+class Transactions
+  attr_accessor :date, :description, :amount, :account_name
+  def initialize(date, description, amount, account_name)
+    @date = date
+    @description = description
+    @amount = amount
+    @account_name = account_name
+  end
+end
