+++ Andriuta/Parser/result.json
@@ -1,1 +1,1 @@
-{"Accounts":[{"name":"BG37FINV91501003939178","balance":"141.02","currency":"BGN","description":"Credit","transactions":[{"date":"28/03/2019","description":"account maintenance charge","amount":"2.00 BGN"},{"date":"13/03/2019","description":"Payment/Cancel POS","amount":"19.68 BGN"},{"date":"13/03/2019","description":"Payment/Cancel POS","amount":"19.68 BGN"},{"date":"13/03/2019","description":"Payment/Cancel POS","amount":"19.68 BGN"},{"date":"12/03/2019","description":"Payment/Cancel POS","amount":"19.68 BGN"}]},{"name":"BG35FINV91502003939180","balance":"10.39","currency":"USD","description":"Debit","transactions":[]}]}
+"{\"Accounts\":[{\"name\":\"BG10FINV91501003939179\",\"balance\":\"EUR\",\"currency\":\"0.00\",\"description\":\"Credit\",\"transactions\":[]},{\"name\":\"BG35FINV91502003939180\",\"balance\":\"USD\",\"currency\":\"10.39\",\"description\":\"Debit\",\"transactions\":[{\"date\":\"06/04/2019\",\"description\":\"Rollover term deposit 032ND50190960001\",\"amount\":\"10.39 USD\"},{\"date\":\"06/04/2019\",\"description\":\"Rollover term deposit 032ND50190960001\",\"amount\":\"10.39 USD\"}]}]}"
